+++ game/interface/enginefunction.cpp
@@ -31,7 +31,7 @@ game::interface::EngineFunction::get(interpreter::Arguments& args)
     // ex int/if/specif.h:IFEngineGet
     int32_t id;
     args.checkArgumentCount(1);
-    if (!interpreter::checkIntegerArg(id, args.getNext(), 1, getDimension(1)-1)) {
+    if (!interpreter::checkIntegerArg(id, args.getNext(), 1, int32_t(getDimension(1)-1))) {
         return 0;
     }
 
@@ -45,14 +45,14 @@ game::interface::EngineFunction::set(interpreter::Arguments& args, const afl::da
 }
 
 // CallableValue:
-int32_t
-game::interface::EngineFunction::getDimension(int32_t which) const
+size_t
+game::interface::EngineFunction::getDimension(size_t which) const
 {
     // ex int/if/specif.h:IFEngineDim
     return (which == 0
             ? 1
             : (m_session.getShipList().get() != 0
-               ? m_session.getShipList()->engines().size()+1
+               ? size_t(m_session.getShipList()->engines().size()+1)
                : 0));
 }
