+++ interpreter/callablevalue.hpp
@@ -40,7 +40,7 @@ namespace interpreter {
         /** Array reflection. Implementation of the IsArray() and Dim() builtins.
             \param which 0=get number of dimensions, 1..n=get that dimension
             \return result */
-        virtual int32_t getDimension(int32_t which) const = 0;
+        virtual size_t getDimension(size_t which) const = 0;
 
         /** Get context for first item in iteration.
             This should be equivalent to <tt>dynamic_cast<Context*>(call(proc, args, true))</tt>,
