+++ test/game/interface/labelvectortest.cpp
@@ -53,7 +53,7 @@ namespace {
             }
         virtual bool isProcedureCall() const
             { return false; }
-        virtual int32_t getDimension(int32_t /*which*/) const
+        virtual size_t getDimension(size_t /*which*/) const
             { return 0; }
         virtual interpreter::Context* makeFirstContext()
             { return 0; }
