+++ test/interpreter/callablevaluetest.cpp
@@ -17,7 +17,7 @@ AFL_TEST("interpreter.CallableValue", a)
             { }
         virtual bool isProcedureCall() const
             { return false; }
-        virtual int32_t getDimension(int32_t /*which*/) const
+        virtual size_t getDimension(size_t /*which*/) const
             { return 0; }
         virtual interpreter::Context* makeFirstContext()
             { return rejectFirstContext(); }
