+++ test/interpreter/processtest.cpp
@@ -211,7 +211,7 @@ namespace {
             }
         virtual bool isProcedureCall() const
             { return m_isProcedure; }
-        virtual int32_t getDimension(int32_t /*which*/) const
+        virtual size_t getDimension(size_t /*which*/) const
             { return 0; }
         virtual interpreter::Context* makeFirstContext()
             { return 0; }
@@ -244,7 +244,7 @@ namespace {
                 args.checkArgumentCount(m_numArgs);
                 m_value = interpreter::toString(value, false);
             }
-        virtual int32_t getDimension(int32_t /*which*/) const
+        virtual size_t getDimension(size_t /*which*/) const
             { return 0; }
         virtual interpreter::Context* makeFirstContext()
             { return 0; }
